Use u32 for AllocId
32 bits ought to be enough for anybody

+++ src/gpu/compute.rs
@@ -37,7 +37,7 @@ pub struct GPUCompute<B: Backend> {
     fence: ManuallyDrop<B::Fence>,
 }
 
-type AllocId = u16;
+type AllocId = u32;
 
 #[derive(Debug, Clone)]
 struct Chunk {
